more flat
middleware folder deleted and became part of index file

because it is a middleware itself
