--- app/model/proto/index.js
@@ -1,4 +1,0 @@
-module.exports = {
-	reqProto : require('./request'),
-	resProto : require('./response')
-}
