--- app/model/proto/response.js
@@ -1,6 +1,0 @@
-const send = require('./sendModel')
-
-module.exports = function (paramObject) {
-	paramObject.send = send
-	return paramObject
-}
