--- app/model/proto/sendModel.js
@@ -1,4 +1,0 @@
-module.exports = function (code, model) {
-	this.status(code).json(model).send().end()
-	return true
-}
