+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "juv",
-  "version": "0.3.17",
+  "version": "0.4.0",
   "description": "Javascript Universal Validation",
   "type": "module",
   "main": "index.js",
