--- test/middleware/index.test.js
@@ -1,9 +1,0 @@
-/* eslint-disable no-undef */
-const app = require('../../app/middleware')
-
-test('should return a function', () => {
-	expect(typeof app).toBe('function')
-})
-
-test('should append the models', () => {
-})
