--- test/model/index.test.js
@@ -1,9 +1,0 @@
-/* eslint-disable no-undef */
-const {reqModel, resModel} = require('../../app/model')
-
-jest.mock('../../app/model')
-
-test('should be a module', () => {
-	expect(typeof reqModel).toBe('object')	
-	expect(typeof resModel).toBe('object')
-})
