+++ test/validations/validateReqModel.test.js
@@ -1,5 +1,5 @@
 /* eslint-disable no-undef */
-const app = require('../../app/validate/validateResModel')
+const app = require('../../validations/validateReqModel')
 
 test('should be a function', () => {
 	expect(typeof app).toBe('function')
