+++ test/validations/validateResModel.test.js
@@ -1,5 +1,5 @@
 /* eslint-disable no-undef */
-const app = require('../../app/validate/validateReqModel')
+const app = require('../../validations/validateResModel')
 
 test('should be a function', () => {
 	expect(typeof app).toBe('function')
